Refactor rename QueryParams to CourtCaseQueryParams
