+++ src/useCases/listCourtCases.ts
@@ -1,11 +1,11 @@
 import { DataSource } from "typeorm"
-import { QueryParams } from "types/QueryParams"
+import { CaseListQueryParams } from "types/CaseListQueryParams"
 import CourtCase from "../entities/CourtCase"
 import getColumnName from "../lib/getColumnName"
 import PromiseResult from "../types/PromiseResult"
 import courtCasesByVisibleForcesQuery from "./queries/courtCasesByVisibleForcesQuery"
 
-const listCourtCases = async (connection: DataSource, queryParams: QueryParams): PromiseResult<CourtCase[]> => {
+const listCourtCases = async (connection: DataSource, queryParams: CaseListQueryParams): PromiseResult<CourtCase[]> => {
   const courtCaseRepository = connection.getRepository(CourtCase)
   const query = courtCasesByVisibleForcesQuery(courtCaseRepository, queryParams.forces)
     .leftJoinAndSelect("courtCase.triggers", "trigger")
@@ -16,8 +16,8 @@ const listCourtCases = async (connection: DataSource, queryParams: QueryParams):
     .limit(queryParams.limit)
 
   if (queryParams.defendantName) {
-    query.where("lower(courtCase.defendantName) like '%' || :name || '%'", {
-      name: queryParams.defendantName.toLowerCase()
+    query.where("courtCase.defendantName ilike '%' || :name || '%'", {
+      name: queryParams.defendantName
     })
   }
